Setup the project as a package module
This commit improves the robustness of the project by packaging it as a python module.

I added the setup.cfg, setup.py and toml file to configure the project appropiately and be detected as a module inside the virtual environment.

I also added mypy and flake8 to improve code readability and detect and logical errors for the src package.

These development tools help to automate the testing process of the project. Now, the pytest module automatically executes with the --cov module to generate the coverage report of the tests.